♻️ Ref(dnd): Separative dnd context as provider from page
